# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Auto Assessor API
  description: Vehicle damage assessment service powered by AI vision
  contact:
    name: Nedbank Avo
    email: avo@nedbank.co.za
  version: 1.0.0
servers:
- url: http://localhost:8085
  description: Generated server url
...
paths:
  /api/admin/assessments/{assessmentId}:
    get:
      tags:
      - Admin Assessment Management
      summary: Get assessment details
      description: "Retrieves a single assessment summary for admin workflows, including\
        \ status, image counters, token usage, and timestamps."
      operationId: getAssessmentDetails
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Assessment summary returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
    put:
      tags:
      - Admin Assessment Management
      summary: Update assessment
      description: Updates editable assessment data such as vehicle details and odometer
        reading.
      operationId: updateAssessment
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssessmentRequest"
        required: true
      responses:
        "200":
          description: Assessment updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentSummaryResponse"
    delete:
      tags:
      - Admin Assessment Management
      summary: Archive assessment
      description: Soft deletes an assessment by marking it as archived. The record
        remains available for restore operations.
      operationId: deleteAssessment
      parameters:
      - name: assessmentId
        in: path
        description: Unique assessment identifier
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Assessment archived successfully
        "404":
          description: Assessment not found
  /api/assessments:
    get:
      tags:
      - Assessments
      summary: List assessments
      description: Returns paginated assessments ordered by pageable input. Each list
        item includes a FRONT_DRIVER_CORNER preview image when available.
      operationId: getAssessments
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Assessments returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAssessmentListItemResponse"
        "400":
          description: Bad request (invalid paging parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAssessmentListItemResponse"
    post:
      tags:
      - Assessments
      summary: Create assessment
      description: Creates a new vehicle assessment by extracting details from a license
        disc image
      operationId: createAssessment
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                licenseDiscImage:
                  type: string
                  format: binary
                  description: License disc image used to extract vehicle details
              required:
              - licenseDiscImage
      responses:
        "201":
          description: Assessment created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAssessmentResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAssessmentResponse"
        "502":
          description: OpenAI integration failure
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAssessmentResponse"
  /api/assessments/{assessmentId}/images:
    post:
      tags:
      - Assessments
      summary: Upload and analyze vehicle image
      description: "Uploads one vehicle image for the requested angle, runs damage\
        \ analysis, and returns findings plus stored image locations."
      operationId: analyzeVehicleImage
      parameters:
      - name: assessmentId
        in: path
        description: Assessment identifier
        required: true
        schema:
          type: string
      - name: angle
        in: query
        description: Camera angle represented by the uploaded image
        required: true
        schema:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Vehicle image file
              required:
              - image
      responses:
        "201":
          description: Image analyzed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentImageUploadResponse"
        "400":
          description: Bad request (validation or illegal argument)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentImageUploadResponse"
        "404":
          description: Assessment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentImageUploadResponse"
        "409":
          description: Duplicate image angle for this assessment
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentImageUploadResponse"
        "502":
          description: OpenAI integration failure
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentImageUploadResponse"
...
  /api/assessments/{assessmentId}:
    get:
      tags:
      - Assessments
      summary: Get assessment by ID
      description: "Returns a full assessment view including vehicle details, all\
        \ analyzed images, and generated report/quote paths."
      operationId: getAssessmentById
      parameters:
      - name: assessmentId
        in: path
        description: ID of the assessment
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Assessment returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetailsResponse"
        "404":
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetailsResponse"
  /api/assessments/{assessmentId}/report.pdf:
    get:
      tags:
      - Assessments
      summary: Download report PDF
      description: Downloads the generated vehicle assessment report PDF for the selected
        assessment.
      operationId: downloadAssessmentReportPdf
      parameters:
      - name: assessmentId
        in: path
        description: ID of the assessment
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Report downloaded successfully
          content:
            application/pdf:
              schema:
                type: string
                format: byte
        "404":
          description: Assessment or generated report not found
          content:
            application/pdf:
              schema:
                type: string
                format: byte
  /api/assessments/{assessmentId}/quote.pdf:
    get:
      tags:
      - Assessments
      summary: Download quote PDF
      description: Downloads the generated vehicle replacement quote PDF for the selected
        assessment.
      operationId: downloadAssessmentQuotePdf
      parameters:
      - name: assessmentId
        in: path
        description: ID of the assessment
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Quote downloaded successfully
          content:
            application/pdf:
              schema:
                type: string
                format: byte
        "404":
          description: Assessment or generated quote not found
          content:
            application/pdf:
              schema:
                type: string
                format: byte
...
components:
  schemas:
    UpdateAssessmentRequest:
      type: object
      properties:
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        odometerReading:
          type: string
        notes:
          type: string
      required:
      - vehicleDetails
    VehicleDetails:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        modelYear:
          type: string
        colour:
          type: string
        engineNumber:
          type: string
        vinNumber:
          type: string
        vehicleCategory:
          type: string
        expiryDate:
          type: string
        extractedFields:
          type: object
          additionalProperties:
            type: string
    AssessmentSummaryResponse:
      type: object
      properties:
        assessmentId:
          type: string
        status:
          type: string
          enum:
          - CREATED
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          - ARCHIVED
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        totalImages:
          type: integer
          format: int32
        successfulImages:
          type: integer
          format: int32
        failedImages:
          type: integer
          format: int32
        totalTokensUsed:
          type: integer
          format: int64
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAssessmentResponse:
      type: object
      properties:
        assessmentId:
          type: string
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        licenseDiscImagePath:
          type: string
        licenseDiscImageContentType:
          type: string
        reportHtmlPath:
          type: string
        reportPdfPath:
          type: string
        quoteHtmlPath:
          type: string
        quotePdfPath:
          type: string
        totalTokensUsed:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    AssessmentImageUploadResponse:
      type: object
      properties:
        assessmentId:
          type: string
        angle:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
        odometerReading:
          type: string
        totalTokensUsed:
          type: integer
          format: int64
        redactionApplied:
          type: boolean
        findings:
          type: array
          items:
            $ref: "#/components/schemas/VehicleDamageFinding"
        originalImagePath:
          type: string
        vehicleImagePath:
          type: string
        vehicleImageContentType:
          type: string
        uploadedAt:
          type: string
          format: date-time
    VehicleDamageFinding:
      type: object
      properties:
        category:
          type: string
        severity:
          type: string
        summary:
          type: string
        panel:
          type: string
        recommendedAction:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    AssessmentListItemResponse:
      type: object
      properties:
        assessmentId:
          type: string
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        odometerReading:
          type: string
        totalTokensUsed:
          type: integer
          format: int64
        reportHtmlPath:
          type: string
        reportPdfPath:
          type: string
        quoteHtmlPath:
          type: string
        quotePdfPath:
          type: string
        frontDriverCornerVehicleImagePath:
          type: string
        frontDriverCornerVehicleImageContentType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PageAssessmentListItemResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentListItemResponse"
        number:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    AssessmentDetailsResponse:
      type: object
      properties:
        assessmentId:
          type: string
        vehicleDetails:
          $ref: "#/components/schemas/VehicleDetails"
        odometerReading:
          type: string
        totalTokensUsed:
          type: integer
          format: int64
        licenseDiscImagePath:
          type: string
        licenseDiscImageContentType:
          type: string
        reportHtmlPath:
          type: string
        reportPdfPath:
          type: string
        quoteHtmlPath:
          type: string
        quotePdfPath:
          type: string
        vehicleImages:
          type: array
          items:
            $ref: "#/components/schemas/VehicleImageResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehicleImageResponse:
      type: object
      properties:
        angle:
          type: string
          enum:
          - FRONT_PASSENGER_CORNER
          - FRONT_VIEW
          - FRONT_DRIVER_CORNER
          - FRONT_DRIVER_SIDE_DOOR
          - REAR_DRIVER_SIDE_DOOR
          - REAR_DRIVER_CORNER
          - REAR_VIEW
          - REAR_PASSENGER_CORNER
          - REAR_PASSENGER_SIDE_DOOR
          - FRONT_PASSENGER_SIDE_DOOR
          - ODOMETER
          - WINDSCREEN_FROM_INSIDE
        redactionApplied:
          type: boolean
        tokensUsed:
          type: integer
          format: int64
        vehicleImagePath:
          type: string
        vehicleImageContentType:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/VehicleDamageFinding"
        uploadedAt:
          type: string
          format: date-time
...